% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_score.R
\name{ks_score}
\alias{ks_score}
\title{KS score}
\usage{
ks_score(
  data,
  y,
  sample_col,
  model = NULL,
  score = NULL,
  return_data = TRUE,
  return_conc_max = FALSE,
  ks_precision = TRUE,
  quebras = NULL,
  score_var_name = "CON_stepKS_score",
  return_ks_num = NULL,
  warn = NULL,
  subsample_data = 1,
  seed = NULL,
  train_sample = "DES"
)
}
\arguments{
\item{data}{R data frame object.}

\item{y}{Name of you target column.}

\item{sample_col}{Name of you sample column.}

\item{model}{R model object.}

\item{score}{Name of you score column.}

\item{return_data}{If TRUE function will return the scored dataset.}

\item{return_conc_max}{If TRUE function will return the conc_max.}

\item{ks_precision}{If TRUE function will score your dataset before compute the KS Score.}

\item{score_var_name}{Name desired of computed score column.}

\item{return_ks_num}{Name of you sample that the function you return the KS Score numeric.}

\item{warn}{Indicate if function will show warnings.}

\item{subsample_data}{Proportion of your data that function will run.}

\item{seed}{Seed to subsample rows.#' @param train_sample Name of you sample that you trained the model.}
}
\value{
scored dataset and the KS score.
}
\description{
This function score (0-1000) a dataset with a binary
classification model, also compute the KS score of the model.
}
