% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sig.R
\name{model_sig}
\alias{model_sig}
\title{model_sig}
\usage{
model_sig(
  model,
  one_cat_sig = FALSE,
  p_value = 0.05,
  ignore_intercept_sig = TRUE,
  trim_chars = 2
)
}
\arguments{
\item{model}{R model object.}

\item{p_value}{p_value.}

\item{ignore_intercept_sig}{If TRUE function will ignore intercept significance.}
}
\value{
TRUE or FALSE
}
\description{
This function return TRUE if you model have all
coefficients significatives and FALSE otherwise.
}
