% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepKS.R
\name{stepKS}
\alias{stepKS}
\title{stepKS}
\usage{
stepKS(
  data,
  y,
  sample_col,
  val_sample,
  train_sample = "DES",
  exclude = NULL,
  include = NULL,
  start = NULL,
  force_in_model = NULL,
  return_type = "none",
  sig_mode = "all",
  direction = "both_sophisticated",
  vars_enable_both = 5,
  near_sample = NULL,
  pct_ks_dif = 0.05,
  flag_bad = 1,
  steps_ahead = 5,
  max_cat = 10,
  ks_precision = FALSE,
  progress_bar = TRUE,
  show_time_elapsed = TRUE,
  ignore_intercept_sig = TRUE,
  p_value = 0.05,
  trim_chars = 2,
  subsample_columns = 1,
  warn = NULL,
  subsample_rows = 1,
  seed = NULL,
  debug = FALSE,
  debug2 = FALSE
)
}
\arguments{
\item{data}{R data frame object.}

\item{y}{Name of you target column.}

\item{sample_col}{Name of you sample column.}

\item{train_sample}{Name of you sample that you trained the model.}

\item{exclude}{If you do not want to pass the name of the Data Frame variables to be analyzed by the algorithm, you also have to pass the list of data frame variables that are not analyzed, for example the key variables, and everything else will be analyzed.}

\item{include}{List with name of variables to be analyzed by the algorithm.}

\item{start}{List of variables that will already begin in the model.}

\item{return_type}{"none","vars","model","formula","scored_data","ahead_vars" or "ahead_model".}

\item{sig_mode}{'off', 'one_cat' or 'all'.}

\item{direction}{"both", "forward" or "both_sophisticated".}

\item{vars_enable_both}{Number of vars in model to enble both method.}

\item{near_sample}{Defines a second sample in which the algorithm will always keep close to the main sample. The sample will need to be in the Data Frame.}

\item{pct_ks_dif}{"Defines how close the KS to the ks_comp need to be close to the ks_focus.}

\item{steps_ahead}{If in the current step when adding the best variable it does not get a KS higher than the KS of the previous step, this parameter indicates how many steps the algorithm will walk forward of the step with the best KS to try to find some variable that increases the KS.}

\item{ks_precision}{If TRUE function will score your dataset before compute the KS Score, for ervery model tested.}

\item{progress_bar}{If TRUE function will show a progress_bar in every step.}

\item{show_time_elapsed}{If TRUE function will show the time_elapsed in every step.}

\item{ignore_intercept_sig}{If TRUE function will ignore intercept significance.}

\item{p_value}{p_value.}

\item{subsample_columns}{Proportion of your data that function will run.}

\item{warn}{Indicate if function will show warnings.}

\item{subsample_rows}{Proportion of your data that function will run.}

\item{seed}{Seed to subsample rows.}
}
\value{
A 'glm' model with the best pool of vars.
}
\description{
This function select the best pool of variables of logistic regression
models based on KS metric and significative coefficients.
}
